requirements.txt
keyboard
numpy
pillow
